requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
py4j==0.10.7
PySocks==1.7.1
pyspark==2.4.5
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.8
